requirements.txt
torch>=1.9.0
transformers>=4.10.0
datasets>=1.11.0
paddlepaddle>=2.2.0
paddlenlp>=2.2.0
ultralytics>=8.0.0
pyyaml>=5.1
numpy>=1.19.5
pandas>=1.3.0
pillow>=8.2.0
tqdm>=4.62.0
scikit-learn>=0.24.2
matplotlib>=3.5.0
tensorboard>=2.6.0 